pyproject.toml
[project]
name = "finance-tracker-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "authx>=1.4.3",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.12",
]

[tool.black]
line-length = 79

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"] # Trust Black here
